Move check/hook methods from client to bot
